Big commit
Addition of the number of files to support users determining their own
term to be translated.

This includes definitions.js to give all definitions of words and
updates to intentschema and the LIST_OF_WORDS to support  this.
However this is not complete as index JS code needs to be updated to
support custom words. Also custom slot types are no longer supported by
Alexa SDK so I will have to look into changing that.
Also put icons into a folder.